Don't require trigger to be mouseover for mouseover_move_with_cursor
Although it makes sense to use it with mouseover, using the keyboard
trigger is the least of one's worries when using this option.

In general, it should be used as:

 mouseover_trigger_behavior: mouse
 mouseover_position: beside_cursor
 mouseover_prevent_cursor_overlap: true
 mouseover_close_el_policy: thumbnail

None of these options should be enforced though, because although the
option becomes annoying to use without that setup, it still works.